[FIX] - Split receipts action types into receipts-types.ts so use-server compiles
Next 16 rejects non-async-function exports from a 'use server' module at
build time. `lib/actions/receipts.ts` was exporting RECEIPT_REASON_MAX_LENGTH,
ReceiptActionErrorCode, and ReceiptActionResult alongside the three async
actions, which broke `yarn build`. Vitest does not enforce that rule, so the
regression slipped past lint/tsc/test gates.

Move the constants/types into a directive-free `lib/actions/receipts-types.ts`
sibling and re-point consumers. The action module now exports only the three
server actions; client components and tests pull the shared constants/types
from `receipts-types`.

`yarn build` now passes — this is the gate slice 5 FE should have honored
from the start.
